Ghidra/Features/Pyhidra/src/main/py/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyhidra"
dynamic = ["version", "readme"]
description = "Native CPython for Ghidra"
license = {text = "Apache-2.0"}
requires-python = ">= 3.9"
authors = [
    { name = "DC3", email = "dc3.tsd@us.af.mil" },
]
maintainers = [
    { name = "Ghidra Development Team" },
	{ name = "DC3", email = "dc3.tsd@us.af.mil" },
]
keywords = [
    "ghidra",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "Jpype1>=1.5.0",
]

[project.optional-dependencies]
testing = [
    "pytest",
    "pytest-datadir",
]

[project.scripts]
pyhidra = "pyhidra.__main__:main"

[project.gui-scripts]
pyhidraw = "pyhidra.gui:_gui"

[project.urls]
Repository = "https://github.com/NationalSecurityAgency/ghidra"

[tool.setuptools.dynamic]
version = {attr = "pyhidra.__version__"}
readme = {file = ["README.md"], content-type = "text/markdown"}

[tool.pytest.ini_options]
required_plugins = ["pytest-datadir"]
addopts = "-p no:faulthandler -m \"not plugin\""
markers = ["plugin"]
